requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
markdown==3.5.1
requests==2.31.0
openai==1.3.7
APScheduler==3.10.4
gunicorn==21.2.0
pytz==2023.3